+++ spec/vcloud/core/vm_spec.rb
@@ -9,7 +9,7 @@ module Vcloud
         @vapp_id = 'vapp-4321'
         @vapp_name = 'test-vapp-1'
         @vm_name = 'test-vm-1'
-        @data_dir = File.join(File.dirname(__FILE__), "../data")
+        @data_dir = File.join(File.dirname(__FILE__), "/data")
         @mock_vm_memory_size = 1024
         @mock_metadata = {
             :foo => "bar",